+++ projects/ui-framework/src/lib/form-elements/date-picker/date-input-directive/dateinput.directive.module.ts
@@ -1,9 +1,11 @@
 import { NgModule } from '@angular/core';
 import { DateInputDirective } from './dateinput.directive';
+import { DateParseService } from '../date-parse-service/date-parse.service';
 
 @NgModule({
   declarations: [DateInputDirective],
   imports: [],
   exports: [DateInputDirective],
+  providers: [DateParseService],
 })
 export class DateInputDirectiveModule {}
