+++ projects/ui-framework/src/lib/form-elements/date-picker/date-range-picker/date-range-picker.component.ts
@@ -20,7 +20,7 @@ import {
 } from '../../../services/utils/functional-utils';
 import { BaseDatepickerElement } from '../datepicker.abstract';
 import { MobileService } from '../../../services/utils/mobile.service';
-import { DateParseService } from '../date-parse.service';
+import { DateParseService } from '../date-parse-service/date-parse.service';
 import { DOMhelpers } from '../../../services/html/dom-helpers.service';
 import { WindowRef } from '../../../services/utils/window-ref.service';
 import { DateRangePickerValue } from '../datepicker.interface';
