+++ projects/ui-framework/src/lib/form-elements/date-picker/date-range-picker/date-range-picker.module.ts
@@ -9,10 +9,10 @@ import { IconsModule } from '../../../icons/icons.module';
 import { CommonModule } from '@angular/common';
 import { InputMessageModule } from '../../input-message/input-message.module';
 import { DateRangePickerComponent } from './date-range-picker.component';
-import { DateParseService } from '../date-parse.service';
+import { DateParseService } from '../date-parse-service/date-parse.service';
 import { EventManagerPlugins } from '../../../services/utils/eventManager.plugins';
 import { B_DATE_FORMATS } from '../dateadapter.mock';
-import { DateInputDirectiveModule } from '../dateinput.directive.module';
+import { DateInputDirectiveModule } from '../date-input-directive/dateinput.directive.module';
 
 @NgModule({
   declarations: [DateRangePickerComponent],
